Move lesson topic modal to GrammarView
