+++ client/components/Lessons/LessonLibrary/index.js
@@ -4,14 +4,13 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useIntl, FormattedMessage } from 'react-intl'
 import { List, WindowScroller } from 'react-virtualized'
 import React, { useEffect, useState } from 'react'
-import { Placeholder, Icon, Select, Container, Modal, TabPane, Tab, Checkbox } from 'semantic-ui-react'
+import { Placeholder, Icon, Container } from 'semantic-ui-react'
 import Stepper from '@keyvaluesystems/react-stepper'
 
 import ScrollArrow from 'Components/ScrollArrow'
 import LibraryTabs from 'Components/LibraryTabs'
 import LessonPracticeTopicsHelp from '../LessonPracticeView/LessonPracticeTopicsHelp'
 import LessonPracticeThemeHelp from '../LessonPracticeView/LessonPracticeThemeHelp'
-import Topics from 'Components/Topics'
 
 import ReactSlider from 'react-slider'
 import { Button } from 'react-bootstrap'
@@ -31,10 +30,6 @@ import {
   lessonsTourViewed,
   updateGroupSelect,
   updateLibrarySelect,
-  updateWordAudio,
-  updateChunkAudio,
-  updateChunkContextAudio,
-  updateAudioTask,
 } from 'Utilities/redux/userReducer'
 import styled from 'styled-components'
 import useWindowDimensions from 'Utilities/windowDimensions'
@@ -81,10 +76,6 @@ const LessonList = () => {
     oid: userId,
     has_seen_lesson_tour,
     vocabulary_score,
-    word_audio,
-    chunk_audio,
-    chunk_context_audio,
-    task_audio,
   } = user
   const {
     pending: metaPending,
@@ -110,8 +101,7 @@ const LessonList = () => {
     private: false,
     group: false,
   })
-
-  const [modal, setModal] = useState(false)
+  
   const dispatch = useDispatch()
   const history = useHistory()
 
@@ -392,58 +382,6 @@ const LessonList = () => {
     return dir * multiplier
   })
 
-  const panes = [
-    {
-      menuItem: 'Select grammar topics',
-      render: () => (
-        <TabPane>
-          <Topics
-            topicInstance={{ ...lesson, instancePending: lessonPending }}
-            editable={libraries.private || (currentGroup && currentGroup.is_teaching)}
-            setSelectedTopics={setSelectedTopics}
-            showPerf={libraries.private}
-          />
-        </TabPane>
-      ),
-    },
-    {
-      menuItem: 'Listening exercise settings',
-      render: () => (
-        <TabPane style={{ display: 'flex', justifyContent: 'center' }}>
-          <div style={{ display: 'flex', flexDirection: 'column', gap: '20px', margin: '40px 0' }}>
-            <Checkbox
-              toggle
-              label={<label style={{ fontWeight: '500' }}>Practice Listening</label>}
-              checked={task_audio}
-              onChange={() => dispatch(updateAudioTask(!task_audio))}
-            />
-            <Checkbox
-              label="Listen to words with context"
-              checked={word_audio}
-              onChange={() => dispatch(updateWordAudio(!word_audio))}
-              disabled={!task_audio}
-              style={{ marginLeft: '37px' }}
-            />
-            <Checkbox
-              label="Listen to phrases without context"
-              checked={chunk_audio}
-              onChange={() => dispatch(updateChunkAudio(!chunk_audio))}
-              disabled={!task_audio}
-              style={{ marginLeft: '37px' }}
-            />
-            <Checkbox
-              label="Listen to phrases with context"
-              checked={chunk_context_audio}
-              onChange={() => dispatch(updateChunkContextAudio(!chunk_context_audio))}
-              disabled={!task_audio}
-              style={{ marginLeft: '37px' }}
-            />
-          </div>
-        </TabPane>
-      ),
-    },
-  ]
-
   const setupViewTitle = () => {
     switch (goStep) {
       case 0:
@@ -486,9 +424,6 @@ const LessonList = () => {
 
   return (
     <>
-      <Modal open={modal} onClose={() => setModal(false)} closeOnEscape closeOnDimmerClick>
-        <Tab panes={panes} />
-      </Modal>
       <LibraryTabs
         values={Object.fromEntries(
           Object.entries(libraries).filter(
@@ -545,10 +480,12 @@ const LessonList = () => {
                   {goStep === 1 && lessonVocabularyControls}
                   <GrammarView
                     currentStepIndex={goStep}
-                    setShowGrammarModal={setModal}
                     lessons={lessons}
                     selectedTopicIds={selectedTopicIds}
                     setSelectedTopics={setSelectedTopics}
+                    topicInstance={{ ...lesson, instancePending: lessonPending }}
+                    editable={libraries.private || (currentGroup && currentGroup.is_teaching)}
+                    showPerf={libraries.private}
                   />
                 </div>
                 <div
